Punted on > 64 fields
