+++ build.sbt
@@ -5,7 +5,6 @@ version := "0.1"
 scalaVersion := "2.13.4"
 
 libraryDependencies += "com.lihaoyi" %% "requests" % "0.6.5"
-
 libraryDependencies += "com.lihaoyi" %% "upickle" % "1.2.2"
 libraryDependencies += "com.lihaoyi" %% "utest" % "0.7.5" % "test"
