+++ src/main/scala/org/overbeck/weather/model/Device.scala
@@ -1,0 +1,9 @@
+package org.overbeck.weather.model
+
+import upickle.default.{ReadWriter, macroRW}
+
+case class Device(macAddress: String, lastData: LastData, info: Info)
+
+object Device {
+  implicit val rw: ReadWriter[Device] = macroRW
+}
