+++ src/main/scala/org/overbeck/weather/model/Info.scala
@@ -1,0 +1,9 @@
+package org.overbeck.weather.model
+
+import upickle.default.{ReadWriter, macroRW}
+
+case class Info(name: String, coords: Coords)
+
+object Info {
+  implicit val rw: ReadWriter[Info] = macroRW
+}
